Initial File Load
Added in my original project(Final.py) as it was written in 2018 and my new code(MainFile.py and ClassFile.py).
Matplotlib will be needed for the graphing functionality. Once installed the MainFile.py and Final.py can be ran.
I have the planetary body data hardcoded in the file, I should add in some file ingestion or something but that could be something to try in the future.